+++ src/main/java/cloud/caravana/anonymouse/classifier/Classifier.java
@@ -3,11 +3,9 @@ package cloud.caravana.anonymouse.classifier;
 import cloud.caravana.anonymouse.Classification;
 import cloud.caravana.anonymouse.Configuration;
 import cloud.caravana.anonymouse.PIIClass;
-import cloud.caravana.anonymouse.Setting;
 
 import java.util.Date;
 import java.util.Optional;
-import java.util.logging.Logger;
 import javax.inject.Inject;
 
 @SuppressWarnings("CdiUnproxyableBeanTypesInspection")
@@ -25,11 +23,11 @@ public abstract class Classifier<T> {
 
     public abstract Optional<Classification> classify(T value, String... context);
 
-    public String generateString(String value, int index, String... context){
+    public String generateString(String value, long index, String... context){
         return null;
     }
 
-    public Date generateDate(Date value, int index, String... context){
+    public Date generateDate(Date value, long index, String... context){
         return null;
     }
