+++ src/main/java/cloud/caravana/anonymouse/classifier/EraseClassifier.java
@@ -1,7 +1,6 @@
 package cloud.caravana.anonymouse.classifier;
 
 import static cloud.caravana.anonymouse.PIIClass.Erase;
-import static cloud.caravana.anonymouse.PIIClass.Telephone;
 
 import cloud.caravana.anonymouse.Classification;
 import java.util.Optional;
@@ -16,7 +15,7 @@ public class EraseClassifier extends Classifier<String> {
     }
 
     @Override
-    public String generateString(String columnValue, int index, String... context) {
+    public String generateString(String columnValue, long index, String... context) {
         return "";
     }
